# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |
    API specification for WASAText messaging application.
    Built according to the PDF specification.
  version: "1.0.0"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Use the user identifier returned from doLogin
  schemas:
    User:
      type: object
      description: Representation of a user in the system
      properties:
        identifier:
          type: string
          description: Unique user identifier (UUID hex string)
          example: "abcdef012345"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        name:
          type: string
          description: Username chosen by the user (3-16 characters)
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_-]+$'
          example: "Maria"
        photo:
          type: string
          format: binary
          description: User profile photo in binary format (optional)
    Group:
      type: object
      description: Representation of a group conversation
      properties:
        groupId:
          type: string
          description: Unique group identifier
          example: "group123"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9_-]+$'
        name:
          type: string
          description: Name of the group
          example: "Study Group"
          minLength: 1
          maxLength: 64
          pattern: '^[a-zA-Z0-9 _-]+$'
        photo:
          type: string
          format: binary
          description: Group photo in binary format (optional)
        members:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: '#/components/schemas/User'
          description: List of users who are members of the group
...
    Error:
      type: object
      description: Standard error response object
      properties:
        message:
          type: string
          description: Human-readable explanation of what went wrong
          example: "Username already taken"
          minLength: 1
          maxLength: 256
          pattern: '^[\s\S]+$'
  parameters:
...
    GroupId:
      name: groupId
      in: path
      description: Group identifier
      required: true
      schema:
        type: string
        description: Group identifier value
        minLength: 1
        maxLength: 64
        pattern: '^[a-zA-Z0-9_-]+$'
paths:
...
  /groups:
    post:
      tags: ["group"]
      summary: Create a new group
      description: |
        User can create a new group with any number of other WASAText users.
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        description: Details of the new group and initial members
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Group creation request
              properties:
                name:
                  type: string
                  description: Descriptive name for the group
                  example: "Study Group"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9 _-]+$'
                memberIds:
                  type: array
                  description: Initial list of member identifiers
                  minItems: 1
                  maxItems: 1000
                  items:
                    type: string
                    description: User identifier
                    minLength: 1
                    maxLength: 64
                    pattern: '^[a-zA-Z0-9_-]+$'
                  example: ["abcdef012345", "fedcba987654"]
              required:
                - name
                - memberIds
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: Invalid group data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/members:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: ["group"]
      summary: Add a user to the group
      description: Group members can add other users to the group.
      operationId: addToGroup
      security:
        - bearerAuth: []
      requestBody:
        description: ID of the user to be added
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Add member request
              properties:
                userId:
                  type: string
                  description: Identifier of the user to invite
                  example: "abcdef012345"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9_-]+$'
              required:
                - userId
      responses:
        '201':
          description: User added to group successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (not a group member)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/members/me:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    delete:
      tags: ["group"]
      summary: Leave a group
      description: Users have the option to leave a group at any time.
      operationId: leaveGroup
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Left the group successfully
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found or not a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/name:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    put:
      tags: ["group"]
      summary: Set the group name
      description: Update the name of a group.
      operationId: setGroupName
      security:
        - bearerAuth: []
      requestBody:
        description: The new group name string
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Rename request
              properties:
                name:
                  type: string
                  description: New name for the group
                  example: "New Group Name"
                  minLength: 1
                  maxLength: 64
                  pattern: '^[a-zA-Z0-9 _-]+$'
              required:
                - name
      responses:
        '200':
          description: Group name updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '400':
          description: Invalid name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (not a member)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/photo:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    put:
      tags: ["group"]
      summary: Set the group photo
      description: Upload or update the group photo.
      operationId: setGroupPhoto
      security:
        - bearerAuth: []
      requestBody:
        description: Binary image file
        required: true
        content:
          image/*:
            schema:
              type: string
              format: binary
              description: Image file data
      responses:
        '200':
          description: Group photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  status:
                    type: string
                    description: Operation result status
                    example: "success"
                    minLength: 1
                    maxLength: 16
                    pattern: '^[a-zA-Z]+$'
        '400':
          description: Invalid photo format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Permission denied (not a member)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'